reorg testing backend to match neo4j for consistency and allow for fewer rebase
